restructure types folder to root
