+++ packages/settings/src/schema/index.ts
@@ -3,4 +3,3 @@ export * from './facility';
 export * from './global';
 export * from './validation';
 export * from './utils';
-export * from './types';
