+++ packages/settings/src/schema/utils.ts
@@ -1,5 +1,5 @@
 import { isObject, isUndefined, mapValues, omitBy } from 'lodash';
-import { Setting, SettingsSchema } from './types';
+import { Setting, SettingsSchema } from '../types';
 
 export const isSetting = (value: Setting | SettingsSchema): value is Setting => {
   return value && isObject(value) && 'type' in value && 'defaultValue' in value;
