+++ packages/settings/src/types/Setting.ts
